+++ man-roxygen/DbmsDetails.R
@@ -24,7 +24,7 @@
 #'                           \code{extraSettings}, and \code{oracleDriver} fields are ignored. If
 #'                           \code{user} and \code{password} are not specified, they are assumed to
 #'                           already be included in the connection string.
-#' @param pathToDriver       Path to the JDBC driver JAR files. See \code{\link{downloadJdbcDrivers}} 
+#' @param pathToDriver       Path to a folder containing the JDBC driver JAR files. See \code{\link{downloadJdbcDrivers}} 
 #'                           for instructions on how to download the relevant drivers.
 #'
 #' @section